requirements.txt
python-telegram-bot==22.5
openai>=1.52.0
